move complex logics to models
