+++ backend/models/follow.js
@@ -67,5 +67,96 @@ module.exports = (sequelize, DataTypes) => {
 
     return followingCount;
   };
+
+  Follow.toggleFollow = async (res, id, user_id) => {
+    let follow = false;
+    const check = await Follow.findOne({
+      where: { follower_id: user_id, following_id: id },
+      order: [["createdAt", "DESC"]],
+    });
+
+    if (check && check.flag) {
+      await Follow.update({ flag: false }, { where: { id: check.id } });
+      await sequelize.models.ActivityLog.create({
+        relatable_id: check.id,
+        relatable_type: "unfollows",
+        user_id,
+      });
+      follow = false;
+    } else if (check && check.flag === false) {
+      await Follow.update({ flag: true }, { where: { id: check.id } });
+      await sequelize.models.ActivityLog.create({
+        relatable_id: check.id,
+        relatable_type: "follows",
+        user_id,
+      });
+      follow = true;
+    } else {
+      const follow = await Follow.create({
+        follower_id: user_id,
+        following_id: id,
+        flag: true,
+      });
+      await sequelize.models.ActivityLog.create({
+        relatable_id: follow.id,
+        relatable_type: "follows",
+        user_id: user_id,
+      });
+      follow = true;
+    }
+
+    return follow;
+  };
+
+  Follow.getFollowing = async (followingsList, followingIdList, id) => {
+    const { Follows } = await sequelize.models.User.findByPk(id, {
+      include: [{ model: Follow, attributes: ["following_id", "flag"] }],
+    });
+
+    await Promise.all(
+      Follows.map(async (follow) => {
+        if (follow.flag) {
+          const { id, first_name, last_name, email, avatar_url, deletedAt } =
+            await sequelize.models.User.findByPk(follow.following_id, {
+              paranoid: false,
+            });
+          if (deletedAt) {
+            return;
+          }
+          followingsList.push({
+            id,
+            first_name,
+            last_name,
+            email,
+            avatar_url,
+            follows: followingIdList.includes(id),
+          });
+        }
+      })
+    );
+  };
+
+  Follow.getFollowers = async (followersList, followingIdList, id) => {
+    const follower = await Follow.findAll({
+      where: { following_id: id, flag: true },
+      attributes: ["follower_id"],
+    });
+
+    await Promise.all(
+      follower.map(async (m) => {
+        const { id, first_name, last_name, email, avatar_url } =
+          await sequelize.models.User.findByPk(m.follower_id);
+        followersList.push({
+          id,
+          first_name,
+          last_name,
+          email,
+          avatar_url,
+          follows: followingIdList.includes(id),
+        });
+      })
+    );
+  };
+
   return Follow;
 };
